Major const refactor
