+++ screenlogicpy/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "0.8.2"
 # flake8: noqa F401
 from screenlogicpy.gateway import ScreenLogicGateway
-from screenlogicpy.const import ScreenLogicError, ScreenLogicRequestError
+from screenlogicpy.const.common import ScreenLogicError, ScreenLogicRequestError
 from screenlogicpy.discovery import async_discover
