+++ screenlogicpy/const/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa F401
+from .common import SLIntEnum, ScreenLogicError, ScreenLogicRequestError
