+++ screenlogicpy/device_const/scg.py
@@ -1,0 +1,4 @@
+from .system import BODY_TYPE
+
+LIMIT_FOR_BODY = {BODY_TYPE.POOL: 100, BODY_TYPE.SPA: 100}
+MAX_SC_RUNTIME = 72
