+++ screenlogicpy/requests/client.py
@@ -1,6 +1,7 @@
 import struct
 
-from ..const import CODE, CLIENT_ID, MESSAGE
+from ..const.common import CLIENT_ID
+from ..const.msg import CODE, COM_MAX_RETRIES
 from .protocol import ScreenLogicProtocol
 from .request import async_make_request
 
@@ -8,7 +9,7 @@ from .request import async_make_request
 async def async_request_add_client(
     protocol: ScreenLogicProtocol,
     clientID: int = CLIENT_ID,
-    max_retries: int = MESSAGE.COM_MAX_RETRIES,
+    max_retries: int = COM_MAX_RETRIES,
 ) -> bool:
     return (
         await async_make_request(
@@ -24,7 +25,7 @@ async def async_request_add_client(
 async def async_request_remove_client(
     protocol: ScreenLogicProtocol,
     clientID: int = CLIENT_ID,
-    max_retries: int = MESSAGE.COM_MAX_RETRIES,
+    max_retries: int = COM_MAX_RETRIES,
 ) -> bool:
     return (
         await async_make_request(
