+++ screenlogicpy/requests/equipment.py
@@ -1,6 +1,6 @@
 import struct
 
-from ..const import CODE
+from ..const.msg import CODE
 from .protocol import ScreenLogicProtocol
 from .request import async_make_request
 from .utility import getSome, getArray
